Refactor.
Remove the very suspicious push back to shapes vector multiple times
when initializing a shape or when creating an anonymous shape. Before
this if 3 shapes were initialized the size of the shapes vector would be
something like 5. Now there should be no problem.
Also there is no need to call draw on anon_shapes vector since anon_shapes
does not actually contain shapes but pointers to shapes so prior to this
for each new anonymous shape there would be a wasted draw call.